Amendment to README.md and Python Learning
A new sub-project is incorporated. Within the sub-project nothing is uploaded yet. A README.md file is added to each of the folders within, ready for further editing. A new section on the general README.md file is added stating important notes and notifications in the form of a to do list (of which can be subject to change).
Potential Harm of Altercation: Extremely Low if Any.